refactor(chen): split API modules
